check for over/underflows in rate_matching added perfect CE to dlsim (compile with PERFECT_CE=1) added visualization to dlsim (compile with XFORMS=1) fixed bug in random_channel for models Rayleigh1_corr and Rayleigh1_anticorr PSS and SSS are now transmitted on one antenna port only, otherwise problems with Rayleigh1_anticorr channel model

+++ trunk/openair1/PHY/impl_defs_lte.h
@@ -521,7 +521,7 @@ typedef struct {
   /// estimated frequency offset (in radians) for all subcarriers
   s32 freq_offset;
   /// eNb_id user is synched to          
-  u8 eNb_id;     
+  s32 eNb_id;     
 } LTE_UE_COMMON;
 
 typedef struct {
